*** empty log message ***
git-svn-id: http://svn.neuromancy.net/magick/trunk/Magick-II@466 1d68a2a4-f605-0410-a2a0-863aa388aa54

+++ src/datetime.cpp
@@ -735,8 +735,8 @@ mstring mDateTime::Ago(bool call)
 {
     mstring Result="";
     int Seconds = SecondsSince();
-    int Years = Seconds / (60 * 60 * 24 * 365.25);
-    Seconds -= Years * 60 * 60 * 24 * 365.25;
+    int Years = (int) (Seconds / (60 * 60 * 24 * 365.25));
+    Seconds -= (int) (Years * 60 * 60 * 24 * 365.25);
     int Days = Seconds / (60 * 60 * 24);
     Seconds -= Days * 60 * 60 * 24;
     int Hours = Seconds / (60 * 60);
